+++ website/forms/forms.py
@@ -1,6 +1,5 @@
 from django.contrib.auth.models import User
 from django import forms
-from website.models.models import Product
 
 class UserForm(forms.ModelForm):
     password = forms.CharField(widget=forms.PasswordInput())
@@ -9,8 +8,3 @@ class UserForm(forms.ModelForm):
         model = User
         fields = ('username', 'email', 'password', 'first_name', 'last_name',)
 
-class ProductForm(forms.ModelForm):
-
-    class Meta:
-        model = Product
-        fields = ('title', 'description', 'price', 'quantity',)
